Fix for static analysis issues
Preventing possible memory leak in function benchmark_run in bench_storage.c
Preventing from bitwise operation on operands of different size
Adding an assertion for NULL

+++ src/datapool/datapool_pmem.c
@@ -141,14 +141,14 @@ datapool_initialize(struct datapool *pool, const char *user_name)
 }
 
 static void
-datapool_flag_set(struct datapool *pool, int flag)
+datapool_flag_set(struct datapool *pool, uint64_t flag)
 {
     pool->hdr->flags |= flag;
     datapool_sync_hdr(pool);
 }
 
 static void
-datapool_flag_clear(struct datapool *pool, int flag)
+datapool_flag_clear(struct datapool *pool, uint64_t flag)
 {
     pool->hdr->flags &= ~flag;
     datapool_sync_hdr(pool);
